requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
crispy-bootstrap4==2024.1
cycler==0.12.1
Django==5.1
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-js-asset==2.2.0
django-tempus-dominus==5.1.2.17
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
